pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "crypto-prediction"
version = "0.1.0"
description = "Multi-model cryptocurrency price prediction and anomaly detection platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "crypto-prediction contributors"},
]
keywords = ["cryptocurrency", "prediction", "deep-learning", "time-series", "anomaly-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Deep learning
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",

    # Exchange connectivity
    "ccxt>=4.2.0",
    "websocket-client>=1.7.0",

    # Technical analysis
    "pandas-ta>=0.3.14b",

    # Classical ML and statistics
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "statsmodels>=0.14.0",

    # Dashboard and visualization
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",

    # Configuration and serialization
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Repository = "https://github.com/your-org/crypto-prediction"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests that take a long time to run",
    "integration: marks integration tests requiring network access",
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
